requirements.txt
asgiref==3.5.2
attrs==22.1.0
Django==4.1.2
django-debug-toolbar==3.7.0
django-filter==22.1
#django-print-sql==2018.3.6
django-seed==0.3.1
djangorestframework==3.14.0
drf-nested-routers==0.93.4
exceptiongroup==1.0.0rc9
Faker==15.1.1
iniconfig==1.1.1
packaging==21.3
pluggy==1.0.0
psycopg2==2.9.4
pyparsing==3.0.9
pytest==7.2.0
python-dateutil==2.8.2
pytz==2022.5
six==1.16.0
sqlparse==0.4.3
tomli==2.0.1
toposort==1.7
yapf==0.32.0
